requirements.txt
asgiref==3.4.0
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-appconf==1.0.4
django-filter==2.4.0
django-guardian==2.4.0
django-imagekit==4.0.2
django-jet==1.0.8
djangorestframework==3.12.4
greenlet==1.1.0
home==0.0.3
idna==2.10
list-print==1.0.1
Markdown==3.3.4
pilkit==2.0
Pillow==8.2.0
psycopg2==2.9.1
Pygments==2.9.0
pyserial==3.5
pytz==2021.1
requests==2.25.1
rfxcom==0.5.0
six==1.16.0
SQLAlchemy==1.4.20
sqlparse==0.4.1
urllib3==1.26.6
